var name change and rebuild as optional
This PR changes the name of `rebuildImage` more representative name of
`containerZoneHealthCheck`.  It changes the rebuild from required to optional.

Modifications to pipelines for name change.
